[gradle] Don't fail import if compose compiler plugin misconfigured.
